+++ game/src/components/menu/config/default_background/cfg_default_background.tsx
@@ -1,5 +1,5 @@
 import SettingsIcon from '@mui/icons-material/Settings';
-import { Grid2, Stack, Typography, useTheme } from '@mui/material';
+import { Grid, Stack, Typography, useTheme } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import style from './cfg_default_background.module.scss';
 
@@ -8,7 +8,7 @@ const CfgScreenDefaultBackground: React.FC = () => {
 	const { t } = useTranslation(['config']);
 
 	return (
-		<Grid2
+		<Grid
 			container
 			direction='column'
 			style={{ backgroundColor: palette.background.default }}
@@ -21,7 +21,7 @@ const CfgScreenDefaultBackground: React.FC = () => {
 				<Typography variant='h4'>{t('config:default_screen_header')}</Typography>
 				<Typography variant='body1'>{t('config:default_screen_body')}</Typography>
 			</Stack>
-		</Grid2>
+		</Grid>
 	);
 };