+++ game/src/components/menu/config/menu/cfg_menu.tsx
@@ -1,14 +1,14 @@
-import { Grid2 } from '@mui/material';
+import { Grid } from '@mui/material';
 import type React from 'react';
 import CfgScreen from '../screen/cfg_screen';
 
 const CfgMenu: React.FC = () => {
 	return (
-		<Grid2 container style={{ height: '100vh' }} justifyContent='center' alignContent='start'>
-			<Grid2 size={{ xs: 8 }} alignSelf='center' height={'100%'}>
+		<Grid container style={{ height: '100vh' }} justifyContent='center' alignContent='start'>
+			<Grid size={{ xs: 8 }} alignSelf='center' height={'100%'}>
 				<CfgScreen showReturnToMenu />
-			</Grid2>
-		</Grid2>
+			</Grid>
+		</Grid>
 	);
 };